requirements.txt
proxy-tools
pythonnet
pywebview ~= 3.7.2